Completed set 4 problem 32
I swapped out my mean-based algorithm for a median-based one, and added
a progress gauge to the output (which now looks pretty rad if I do say
so myself). This could easily crack my initial goal of 5 ms, so I
lowered the delay time to 1 ms. It still works!

This completes set 4.